don't start template informer unless templateservicebroker is configured
